requirements.txt
torch
torchvision
flask
Pillow
